pyproject.toml
[project]
name = "iwdd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface>=0.0.1",
    "lightning>=2.5.5",
    "numpy>=2.3.4",
    "torchvision>=0.24.0",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "pre-commit>=3.5.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
